+++ src/components/Modal/Login.tsx
@@ -1,0 +1,5 @@
+import { FunctionComponent } from "react";
+
+export const Login: FunctionComponent = () => {
+  return <div>Login</div>;
+};
