+++ src/components/Modal/Register.tsx
@@ -1,0 +1,5 @@
+import { FunctionComponent } from "react";
+
+export const Register: FunctionComponent = () => {
+  return <div>Register</div>;
+};
