+++ src/components/UserProfile/LoggedIn.tsx
@@ -1,0 +1,5 @@
+import { FunctionComponent } from "react";
+
+export const LoggedIn: FunctionComponent = () => {
+  return null;
+};
